+++ Util/PixelPerfect.lua
@@ -22,12 +22,40 @@ local L = _p.L;
 _p.PixelPerfect = {}
 local PixelPerfect = _p.PixelPerfect;
 
+local _scaledUiToPixelMulti;
+
+local function NearestPixelSize(uiUnits)
+    if (uiUnits == 0 or _scaledUiToPixelMulti == 1) then
+       return uiUnits;
+    end
+    
+    local y = _scaledUiToPixelMulti;
+    if (_scaledUiToPixelMulti > 1) then
+       --if real pixels are "bigger" than ui units (screenSize / uiScale > physicalSize) tend to make items bigger
+       y = -y;
+    end
+    if (uiUnits < 0) then
+       --invert for positive units to keep the "rounding" behavior the same between negative and positive
+       y = -y;
+    end
+    return uiUnits - uiUnits % y;
+ end
+
+local function CalculateScalingFactors()
+    local _, physHeight = GetPhysicalScreenSize();
+    local _uiToPixelMulti = 768 / physHeight;
+    _scaledUiToPixelMulti = _uiToPixelMulti / UIParent:GetScale();
+
+end
+
 local eventFrame = CreateFrame("Frame");
 eventFrame:SetScript("OnEvent", function(_, event, cvar, _)
     if (event == "UI_SCALE_CHANGED") then
+        CalculateScalingFactors();
     end
 end);
 eventFrame:RegisterEvent("UI_SCALE_CHANGED");
+CalculateScalingFactors();
 
 ---Sets the size respecting the pixel grid.
 ---@param region Frame | Region
@@ -35,25 +63,36 @@ eventFrame:RegisterEvent("UI_SCALE_CHANGED");
 ---@param height number
 function PixelPerfect.SetSize(region, width, height)
     height = height or width;
-    region:SetSize(width, height);
+    region:SetSize(NearestPixelSize(width), NearestPixelSize(height));
 end
 
 ---Sets the width respecting the pixel grid.
 ---@param region Frame | Region
 ---@param width number
 function PixelPerfect.SetWidth(region, width)
-    region:SetWidth(width);
+    region:SetWidth(NearestPixelSize(width));
 end
 
 ---Sets the height respecting the pixel grid.
 ---@param region Frame | Region
 ---@param height number
 function PixelPerfect.SetHeight(region, height)
-    region:SetHeight(height);
+    region:SetHeight(NearestPixelSize(height));
 end
 
 ---Sets the point respecting the pixel grid.
 ---@param region Frame | Region
-function PixelPerfect.SetPoint(region, ...)
-    region:SetPoint(...);
+function PixelPerfect.SetPoint(region, arg1, arg2, arg3, arg4, arg5, ...)
+    if not arg2 then arg2 = region:GetParent() end
+
+    if type(arg2)=='number' then arg2 = NearestPixelSize(arg2) end
+	if type(arg3)=='number' then arg3 = NearestPixelSize(arg3) end
+	if type(arg4)=='number' then arg4 = NearestPixelSize(arg4) end
+	if type(arg5)=='number' then arg5 = NearestPixelSize(arg5) end
+
+    region:SetPoint(arg1, arg2, arg3, arg4, arg5, ...);
+end
+
+function PixelPerfect.GetPixelSize()
+    return _scaledUiToPixelMulti;
 end